+++ Magic/CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Beginner, Apprentice, Student, and Master wands are now for demo purposes only and not tied to player progression.
  - Fix sniper rifle damage not undoing
+ - The /mskills menu will now work with a player's class skills
 
 ## 7.3